third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-ui/click-gutter-breakpoint-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that breakpoints can be added and removed by clicking the gutter.

Script execution paused.
Breakpoint sidebar pane while paused
click-breakpoints.js:4    var b = 1; // The breakpoint should happen here
Script execution resumed.
